requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.0
attrs==25.4.0
captum==0.8.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.1
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.4.1
debugpy==1.8.16
decorator==5.2.1
diffusers==0.35.1
dill==0.4.0
executing==2.2.1
facenet-pytorch==2.6.0
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
grad-cam==1.5.5
grpcio==1.76.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
ImageIO==2.37.2
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lazy_loader==0.4
lightning-utilities==0.15.2
llvmlite==0.45.1
lpips==0.1.4
Markdown==3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.5
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytorch_ssim==0.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
sentry-sdk==2.48.0
shap==0.49.1
six==1.17.0
slicer==0.0.8
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.12.12
torch==2.2.2
torch-fidelity==0.3.0
torchaudio==2.8.0
torchcam==0.4.1
torchmetrics==1.8.2
torchvision==0.17.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transforms==0.2.1
triton==2.2.0
ttach==0.0.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
wandb==0.23.1
wcwidth==0.2.13
Werkzeug==3.1.4
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
torch==2.2.2+cu121
torchvision==0.17.2+cu121
-f https://download.pytorch.org/whl/cu121